app/requirements.txt
flask
flask-cors
sqlalchemy
pydantic
python-multipart
openai
google-auth-oauthlib
google-api-python-client
beautifulsoup4